pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cop-cli"
version = "0.1.0"
description = "Context-Oriented Programming CLI - Build and deploy LLM agents from context modules"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "COP Research Team"}
]
keywords = ["llm", "ai", "context", "prompts", "fine-tuning", "langraph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "jsonschema>=4.17.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "diskcache>=5.6.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
pipeline = [
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "numpy>=1.24.0",
]
all = [
    "cop-cli[dev,pipeline]",
]

[project.scripts]
cop = "cop.cli:main"

[project.urls]
Homepage = "https://github.com/context-oriented-programming/cop-cli"
Documentation = "https://github.com/context-oriented-programming/cop-cli/docs"
Repository = "https://github.com/context-oriented-programming/cop-cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["cop*"]

[tool.setuptools.package-data]
cop = ["schemas/*.json"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

